Introduce a Notification Email mixin that can be used by both Assignment and Quiz editors
